requirements.txt
matplotlib
reportlab
fpdf
mysql-connector-python
Pillow
bcrypt
openpyxl